fix: delete contact by not only email but user id

+++ backend/app/database/crud.py
@@ -106,12 +106,12 @@ def add_contact_list(db: Session, email: str, user_id: int, is_subscribed: bool)
         session.close()
 
 
-def delete_contact_by_email(db: Session, email: str):
+def delete_contact_by_email_and_user_id(db: Session, email: str, user_id: int):
     session = Session()
     try:
-        db.query(models.ContactList).filter(models.ContactList.email == email).delete()
+        db.query(models.ContactList).filter(models.ContactList.email == email, models.ContactList.user_id == user_id).delete()
         db.commit()
-    except Exception as err:
+    except:
         session.rollback()
         raise
     finally:
